Migrate console functions to Console.h (#3626)
* Migrate console functions to Console.h

* include <array>

* apply style
