+++ cpp/open3d/Open3D.h.in
@@ -95,6 +95,7 @@
 #include "open3d/t/pipelines/slac/SLACOptimizer.h"
 #include "open3d/t/pipelines/voxelhashing/Frame.h"
 #include "open3d/t/pipelines/voxelhashing/Model.h"
+#include "open3d/utility/Console.h"
 #include "open3d/utility/Eigen.h"
 #include "open3d/utility/FileSystem.h"
 #include "open3d/utility/Helper.h"
