+++ cpp/open3d/core/SizeVector.h
@@ -28,6 +28,7 @@
 
 #include <cstddef>
 #include <numeric>
+#include <sstream>
 #include <string>
 #include <vector>
