+++ cpp/open3d/t/geometry/TSDFVoxelGrid.cpp
@@ -26,7 +26,6 @@
 
 #include "open3d/t/geometry/TSDFVoxelGrid.h"
 
-#include "open3d/Open3D.h"
 #include "open3d/t/geometry/PointCloud.h"
 #include "open3d/t/geometry/kernel/TSDFVoxelGrid.h"
 #include "open3d/utility/Logging.h"
