+++ cpp/open3d/utility/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_library(utility OBJECT)
 
 target_sources(utility PRIVATE
+    Console.cpp
     Eigen.cpp
     FileSystem.cpp
     Helper.cpp
