+++ cpp/open3d/utility/Logging.h
@@ -26,10 +26,8 @@
 
 #pragma once
 
-#include <Eigen/Core>
-#include <iostream>
+#include <functional>
 #include <string>
-#include <vector>
 
 #ifndef FMT_HEADER_ONLY
 #define FMT_HEADER_ONLY 1
@@ -302,35 +300,6 @@ private:
     size_t progress_pixel_;
     bool active_;
 };
-
-std::string GetCurrentTimeStamp();
-
-std::string GetProgramOptionAsString(int argc,
-                                     char **argv,
-                                     const std::string &option,
-                                     const std::string &default_value = "");
-
-int GetProgramOptionAsInt(int argc,
-                          char **argv,
-                          const std::string &option,
-                          const int default_value = 0);
-
-double GetProgramOptionAsDouble(int argc,
-                                char **argv,
-                                const std::string &option,
-                                const double default_value = 0.0);
-
-Eigen::VectorXd GetProgramOptionAsEigenVectorXd(
-        int argc,
-        char **argv,
-        const std::string &option,
-        const Eigen::VectorXd default_value = Eigen::VectorXd::Zero(0));
-
-bool ProgramOptionExists(int argc, char **argv, const std::string &option);
-
-bool ProgramOptionExistsAny(int argc,
-                            char **argv,
-                            const std::vector<std::string> &options);
 
 }  // namespace utility
 }  // namespace open3d
