+++ cpp/open3d/visualization/visualizer/VisualizerRender.cpp
@@ -29,6 +29,7 @@
 #include "open3d/io/IJsonConvertibleIO.h"
 #include "open3d/io/ImageIO.h"
 #include "open3d/io/PointCloudIO.h"
+#include "open3d/utility/Helper.h"
 #include "open3d/visualization/utility/GLHelper.h"
 #include "open3d/visualization/visualizer/ViewParameters.h"
 #include "open3d/visualization/visualizer/ViewTrajectory.h"
